+++ R/brsmatch.R
@@ -11,9 +11,9 @@
 #'
 #' Note that when using exact matching, the `n_pairs` are split roughly in
 #' proportion to the number of treated subjects in each exact matching group.
-#' This has a possibility of failing  when `n_pairs` is large.  If this happens
-#' to you, we suggest manually performing exact matching, for example with
-#' `split()`, and selecting `n_pairs` for each group interactively.
+#' If you would like to control `n_pairs` exactly, we suggest manually
+#' performing exact matching, for example with `split()`, and selecting
+#' `n_pairs` for each group interactively.
 #'
 #' @param n_pairs The number of pairs desired from matching.
 #' @param data A data.frame or similar containing columns matching the `id,
@@ -121,6 +121,20 @@ brsmatch <- function(
     data[[trt_time]] <- data[[trt_time]] - 1
   }
 
+  id_list <- unique(data[[id]]) # compute before any NA removal
+
+  # Remove NA rows except those in `trt_time` column, with a warning
+  na_action <- stats::na.omit(data[, setdiff(colnames(data), trt_time)])
+  na_rows <- attributes(na_action)$na.action
+  if (!is.null(na_rows)) {
+    rlang::warn(c(
+      "ID, time, and covariates should not have NA entries.",
+      i = paste("Removed", length(na_rows), "rows.")
+    ))
+    data <- data[-na_rows, ]
+  }
+
+
   if (!is.null(exact_match)) {
     data_split <- split(data, data[, exact_match, drop = FALSE])
     covariates <- setdiff(covariates, exact_match)
@@ -151,7 +165,7 @@ brsmatch <- function(
     )
   }
 
-  .output_pairs(matched_ids, id = id, id_list = unique(data[[id]]))
+  .output_pairs(matched_ids, id = id, id_list = id_list)
 }
 
 .brsmatch <- function(
@@ -168,44 +182,30 @@ brsmatch <- function(
   optimizer <- options$optimizer
   verbose <- options$verbose
 
-  if (verbose) {
-    rlang::inform("Computing distance from data...")
+  .print_if <- function(condition, message, ...) {
+    if (condition) {
+      rlang::inform(message, ...)
+    }
   }
+
+  .print_if(verbose, "Computing distance from data...")
   edges <- .compute_distances(data, id, time, trt_time, covariates, options)
 
   bal <- NULL
   if (balance) {
-    if (verbose) {
-      rlang::inform("Building balance columns from data...")
-    }
+    .print_if(verbose, "Building balance columns from data...")
     bal <- .balance_columns(data, id, time, trt_time, balance_covariates)
   }
 
-  if (verbose) {
-    rlang::inform("Constructing optimization model...")
-  }
+  .print_if(verbose, "Constructing optimization model...")
   model <- .rsm_optimization_model(n_pairs, edges, bal, optimizer, verbose, balance)
 
-  if (verbose) {
-    rlang::inform("Preparing to run optimization model")
-  }
-  if (optimizer == "gurobi") {
-    res <- gurobi::gurobi(model, params = list(OutputFlag = 1 * verbose))
-    matches <- res$x[grepl("f", model$varnames)]
-  } else if (optimizer == "glpk") {
-    res <- Rglpk::Rglpk_solve_LP(
-      model$obj,
-      model$mat,
-      model$dir,
-      model$rhs,
-      types = model$types,
-      max = model$max,
-      control = list(verbose = verbose, presolve = TRUE)
-    )
-    # res <- with(model, Rglpk::Rglpk_solve_LP(obj, mat, dir, rhs, types = types, max = max,
-    #                                          control = list(verbose = verbose, presolve = TRUE)))
-    matches <- res$solution[grepl("f", model$varnames)]
-  }
+  .print_if(verbose, "Preparing to run optimization model")
+  res <- .solve_or_reduce_pairs(n_pairs, model, optimizer, verbose)
+  matches <- switch(optimizer,
+    "gurobi" = res$x[grepl("f", model$varnames)],
+    "glpk" = res$solution[grepl("f", model$varnames)]
+  )
 
   matched_ids <- edges[matches == 1, c("trt_id", "all_id"), drop = FALSE]
   return(matched_ids)
@@ -581,4 +581,64 @@ brsmatch <- function(
     ))
   }
   return(model)
+}
+
+
+#' Solve brsmatch model even if too many pairs specified
+#'
+#' If the `n_pairs` is too large, the model will be infeasible, and will return
+#' a status code indicating this. This function will iteratively reduce the
+#' number of pairs until the model becomes solvable, then will return the
+#' solution with a warning.
+#'
+#' NOTE: gurobi functionality is untested, as I have since lost my license.
+#' Code is based on the documentation at
+#' https://www.gurobi.com/documentation/current/refman/r_grb.html
+#'
+#' @inheritParams brsmatch
+#' @param model The model output from `.rsm_optimization_model()`
+#'
+#' @return The result from [Rglpk::Rglpk_solve_LP] after possible n_pair
+#'   reduction.
+#'
+#' @noRd
+.solve_or_reduce_pairs <- function(n_pairs, model, optimizer, verbose) {
+  n_pairs_solve <- n_pairs
+
+  while (TRUE) {
+    if (optimizer == "gurobi") {
+      res <- gurobi::gurobi(model, params = list(OutputFlag = 1 * verbose))
+    } else if (optimizer == "glpk") {
+      res <- Rglpk::Rglpk_solve_LP(
+        model$obj,
+        model$mat,
+        model$dir,
+        model$rhs,
+        types = model$types,
+        max = model$max,
+        control = list(verbose = verbose, presolve = TRUE)
+      )
+    }
+    solved <- switch(optimizer,
+      "gurobi" = res$status == "OPTIMAL",
+      "glpk" = res$status == 0,
+    )
+    if (solved) {
+      break
+    }
+
+    n_pairs_solve <- n_pairs_solve - 1
+    # n pairs only appears in the first two model RHS constraints
+    model$rhs[1:2] <- c(n_pairs_solve, -n_pairs_solve)
+  }
+
+  if (n_pairs_solve != n_pairs) {
+    rlang::warn(
+      paste(
+        "Number of pairs reduced from", n_pairs, "to",
+        n_pairs_solve, "to create a solveable model."
+      )
+    )
+  }
+  return(res)
 }